+++ rmk-macro/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.6] - 2024-06-14
+
 ### Added
 
 - Support more nRF chips: nRF52833, nRF52810, nRF52811